requirements.txt
alembic==1.7.7
black==22.3.0
blinker==1.4
click==8.0.4
colorama==0.4.4
coverage==6.3.2
docformatter==1.4
Flask==2.0.3
flask-cas-ng==1.1.0
Flask-DB==0.3.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
mypy==0.950
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pydocstyle==6.1.1
python-dotenv==0.20.0
six==1.16.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.32
SQLAlchemy-Utils==0.38.2
toml==0.10.2
tomli==2.0.1
types-Flask-SQLAlchemy==2.5.2
types-SQLAlchemy==1.4.34
typing_extensions==4.2.0
untokenize==0.1.1
Werkzeug==2.0.3
xmltodict==0.12.0
